build(helpers): 🚚  Rename to `shared`
