+++ packages/eslint/package.json
@@ -78,7 +78,7 @@
 	"devDependencies": {
 		"@terminal-nerds/typescript-config": "0.3.1",
 		"@types/eslint": "8.4.10",
-		"@workspace/helpers": "workspace:*",
+		"@workspace/shared": "workspace:*",
 		"typescript": "4.9.4"
 	}
 }
