+++ packages/prettier/package.json
@@ -48,7 +48,7 @@
 	"devDependencies": {
 		"@terminal-nerds/typescript-config": "0.3.1",
 		"@types/prettier": "2.7.1",
-		"@workspace/helpers": "workspace:*",
+		"@workspace/shared": "workspace:*",
 		"typescript": "4.9.4"
 	}
 }
