+++ packages/stylelint/package.json
@@ -53,7 +53,7 @@
 	},
 	"devDependencies": {
 		"@terminal-nerds/typescript-config": "0.3.1",
-		"@workspace/helpers": "workspace:*",
+		"@workspace/shared": "workspace:*",
 		"stylelint": "14.16.0",
 		"typescript": "4.9.4"
 	}
