+++ shared/package.json
@@ -2,7 +2,7 @@
 	"$schema": "https://json.schemastore.org/package",
 	"private": true,
 	"type": "module",
-	"name": "@workspace/helpers",
+	"name": "@workspace/shared",
 	"version": "0.0.0",
 	"engines": {
 		"node": ">=18"
