+++ shared/source/configuration.ts
