+++ shared/source/environment.ts
