+++ shared/source/module.ts
